sources/inbox/requirements.txt
pyyaml
markdown
python-docx
